+++ src/main/java/com/product/inventory/service/ProductServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -105,7 +106,7 @@ public class ProductServiceImpl implements ProductService {
     public Map<String, Object> getInventorySummary() {
         long totalProducts = repo.countProducts();
         long totalQuantity = repo.sumQuantities();
-        double averagePrice = repo.averagePrice();
+        BigDecimal averagePrice = repo.averagePrice();
         List<OutOfStockProduct> outOfStock = repo.findByQuantity(0);
 
         return generateSummaryMap(totalProducts, totalQuantity, averagePrice, outOfStock);
@@ -114,7 +115,7 @@ public class ProductServiceImpl implements ProductService {
     /**
      * Creates summary map from given parameters
      */
-    private Map<String, Object> generateSummaryMap(long totalProducts, long totalQuantity, double averagePrice,
+    private Map<String, Object> generateSummaryMap(long totalProducts, long totalQuantity, BigDecimal averagePrice,
                                                    List<OutOfStockProduct> outOfStock) {
         Map<String, Object> summary = new HashMap<>();
         summary.put("totalProducts", totalProducts);
